Removed W_CheckNoAmmo
Optimized how the mod handles the event when running out of ammo while shooting. Checks when pressing button0 and letting off the button instead every frame (resetting the weaponmodel).
If SmartWeapon is enabled then the weapon in hand will switch to a better weapon, like standard Quake. If SmartWeapon is disabled or `w_switch` enabled then the gun with not change automatically.